fix: 🐛 change to long long int

+++ conv_parallel.c
@@ -24,7 +24,7 @@ int main() {
     // printf("Number of threads: %d\n", num_threads);
     omp_set_num_threads(num_threads);
     
-    int *ans = malloc(sizeof(int) * (NA - NF + 1));
+    long long int *ans = malloc(sizeof(long long int) * (NA - NF + 1));
     for (int i = 0; i < NA - NF + 1; i++) {
         ans[i] = 0;
     }
@@ -37,7 +37,7 @@ int main() {
     }
 
     for (int i = 0; i < NA - NF + 1; i++) {
-        printf("%d\n", ans[i]);
+        printf("%lld\n", ans[i]);
     }
 
     // ---- free memory ----
